AES: hdl: aes_axi_stream: swap input and output bytes
The linux crypto API provides a pointer to a 16 byte buffer in memory
(big endian representation of 4 x 32-bit words, as per the AES standard).

When the data is sent on the AXI bus (little endian), the bytes get
swapped.

Therefore, Swap the bytes again in hardware to get the original
big endian representation.

Signed-off-by: Ovidiu Panait <ovpanait@gmail.com>