hdl: Get rid of swapping 32-bit words
The AES standard counts the bytes in a block from left to right (LSB -> MSB).
Using this encoding in HDL introduced a lot of unnecessary complexity. Because
of this, we switch to the normal bit direction (MSB <- LSB) so the byte order
is reversed from the standard. The rest of the AES processing remains the same,
just the order of bytes will be reversed.

Previously, to achive the same byte encoding as in the AES standard, we were
running a swap operation on each 32-bit word after retrieving the data from the
BRAM. (this was needed due to the way the AES blocks are stored in the Linux
kernel buffers). However, this shuffling makes debugging really
confusing. Now that the bit direction has been reversed to MSB <-> LSB, we can
get rid of swapping as well.

Signed-off-by: Ovidiu Panait <ovidiu.panait@windriver.com>