hdl: aes_controller: Pull CBC logic out of the control FSM
This facilitates adding new modes of operation without changing the FSM logic.
A mode of operation should consist of a separate module the dictates how the
plaintext/ciphertext/iv is computed.

Now that the FSM is generic enough, adding new block modes of operation should
hopefully be not as intrusive.

Signed-off-by: Ovidiu Panait <ovpanait@gmail.com>