tb: Add standalone zynq_aes testbench
Currently, the main testbench depends on the AXI VIP infrastructure. This makes
the simulation take a lot of time (due tothe fact that a project needs to be
created and a block design sourced) and also makes finetuning hard.

This commit introduces a standalone testbench for the top module. AXI Stream
master and slave controllers are simulated as well.

Signed-off-by: Ovidiu Panait <ovpanait@gmail.com>