hdl: aes_controller: Fix data overwrite corner case
Currently, if the output FIFO was busy for longer than the time it would take
to compute a new AES block, the block waiting to be stored in the FIFO would be
overwritten by the new one, causing controller hangs.

Ths bug was around for quite a while, but only now, after the pipeline stages
separation, popped up during simulation. So far the testing seems to indicate
that this was the cause for #4. More testing will be performed to confirm this.

NOTE: This is just a temporary workaround, as it introduces some performance
penalty (~2MB/s) due to the fact that the next AES operation won't start until
the last one was stored in the fifo. The proper fix might require a little bit
of rework on the FSM side.

TODO: add formal verificaton testcase for this.

Signed-off-by: Ovidiu Panait <ovpanait@gmail.com>