aes_controller: Introduce op_done wrapper
Let each block cipher mode define when a processing operation is done. The
current behavior is to consider that an operation is complete when the AES
encryption/decryption/key expansion computation is over. This may not be the
case for all modes of operation. For instance, GCM mode needs to perform
multiple operations in one pass (encryption + authentification).

Signed-off-by: Ovidiu Panait <ovpanait@gmail.com>