hdl: aes_controller: Factor aes control signals out of fsm
The current behavior is to perform an aes encrypton/decryption/key expansion
for each key/data block retrieved from the FIFO. However, this is not suitable
for modes of operations such as GCM which require multiple aes crypto
operations at specific stages of the processing (computing the subkey H,
computing the final tag, etc.).

Therefore, we need to factor out the aes control signals from the FSM so that
they can be controlled from outside the FSM, by each mode of operation.

Signed-off-by: Ovidiu Panait <ovpanait@gmail.com>