hdl: aes_controller: Shift processing tasks from controller to module simplmenting modes of operation
Currently, the controller takes care of starting AES operations and decides
which data gets sent on the bus. However, this could work only because all the
modes of operations implemented were similar in essence. However, GCM is
different and more complex because, so the module implementing it needs to get
full control of the AES algorithm siganls and the output bus.

This change lets the modules implementing crypto modes of operations decide how
to process the input blocks. The controller's job is now only to expand the key
and pass all the input blocks to the next pipeline stage. Modules implementing
modes of operation decide how input data should be treated (IV, AAD, etc.),
when to start an AES operation, what data to place on the output bus.

Temporarily remove all modes of operations other than ECB, until they are
converted as well.

Signed-off-by: Ovidiu Panait <ovpanait@gmail.com>