hdl: block_ram: Switch to dual port block RAM
Using a single port block RAM configuration is not worth it because the fifo
must use the same address port for both read and write operations. This
complicates the fifo and decreases throughput.

Such a change is also needed in order to easly convert this synchronous design
to an asynchronous one.

* Switch to dual port BRAM in block_ram module
* Remove the fifo_empty/fifo_full flags from the FIFO
* Use new address pointer scheme to get rid of fifo_has_data logic
* Update formal properties
* Update testbench

Signed-off-by: Ovidiu Panait <ovpanait@gmail.com>