hdl: aes_controller: Use the new FIFO module
The fifo.v module was adjusted to use a dual-port bram implementation. This
simplified the fifo a lot and made the fifo_empty/full/almost_* flags internal.
Also, the fifo can now be only 2^N in depth, so the external DEPTH parameter
was removed. The depth is computed internally based on the ADDRESS_WIDTH
parameter.

Adjust all the aes_controller modules to reflect these changes.

Signed-off-by: Ovidiu Panait <ovpanait@gmail.com>