hdl: fifo: Add output buffer on read side
In order to allow for read transactions on consecutive clock edges, we need to
pre-read the next data from BRAM. In order to do this, incease the read pointer
by 1 after the fifo goes empty -> non-empty, so that bram_r_data will hold the
next BRAM word. This word can either be valid or bogus, depending on the value
of fifo_empty.

r_skidbuf will hold the output fifo read value if a transaction does not take
place on the first clock edge after fifo_read_tvalid is asserted (we must store
this first word somewhere, because we increase the BRAM read
address and it would otherwise be overwritten).

Also, remove some formal properties that no longer apply(e.g. a read
transaction can take place even if the fifo is empty, because
we already stored the BRAM data into the read buffer).

Signed-off-by: Ovidiu Panait <ovpanait@gmail.com>