hdl: Add block ram output register
BRAM clock to out delay is pretty high, making subsequent paths to be really
tight on timing. In order to slightly improve design Fmax, register the read
output of the memory element which stores expanded keys. Introduce a key_valid
signal for both encryption and decryption modules in order to indicate when to
sample the round key.

Signed-off-by: Ovidiu Panait <ovpanait@gmail.com>