hdl: gcm: Register gctr outputs
When the input data (plaintext/ciphertext) size is not a multiple of the GCM
block size, we need to zero out the extra least significant bits. In order to
achieve this, register the gctr output and postprocess it in the next clock.

Update all instances using gctr outputs to use the final values and make sure
not to start a new gctr operation after the last plaintext/ciphertext block is
computed (update crypto_ready state).

Signed-off-by: Ovidiu Panait <ovpanait@gmail.com>