hdl: gcm: Add gcm_mask module
Currently, cumbersome shift logic is used to generate a mask that is applied on
the last payload block (when payload size is not a multiple of 128). This make
the design fail timing with the default 142 MHz clock.

Introduce a slimmed-down gcm_mask module that takes a register representing the
number of bits to be masked to 1and generates an output block accordingly. For
instance:

WIDTH = 16
SEL_WIDTH = 4
sel = 0111

mask -> 1111111000000000

Signed-off-by: Ovidiu Panait <ovpanait@gmail.com>