driver: zynqaes: Introduce tx/rx remainder fields to zynqaes_dma_ctx
Introduce {rx, tx}_remainder fields to zynqaes_dma_ctx struct. We use them to
keep track of the scatterlist resizing operations that we perform if the
scatterlist array contains more bytes than what the HW expects.

Currently, we do not restore the input scatterlist passed from the crypto API
back to its original state, which needs to be fixed in a subsequent commit.

Signed-off-by: Ovidiu Panait <ovpanait@gmail.com>