driver: zynqaes_skcipher: Restore original scatterlists in dma_callback
The HW expects reqctx->nbytes for skcipher operations, so we resize the
scatterlist passed from the crypto API before starting the dma transfer. This
is the case, for instnace, for IPSEC requests that append data at the end of
the last scatterlist (ESP authentification data). However, we do not restore
the scatterlists befeore finalizing the crypto request. Even though this did
not generate errors so far, it should be fixed.

Signed-off-by: Ovidiu Panait <ovpanait@gmail.com>