driver: aead: fix scatterwalk_ffwd() callsite
Currently, we wrongly assume that we can discard the return value of
scatterwalk_ffwd(), because it would be the same pointer we passed via dst
argument. However, scatterwalk_ffwd() can also return a pointer inside the
original scatterlist, causing oopses in subsequent logic.

To fix this, capture the return value of scatterwalk_ffwd() and use it instead.

Signed-off-by: Ovidiu Panait <ovpanait@gmail.com>